feat: Integrating implemented component-by-component logic

+++ src/main/kotlin/lotto/domain/LottoExchanger.kt
@@ -23,9 +23,9 @@ class LottoExchanger(val purchaseAmount: Int) {
         }
     }
 
-    fun yieldRate(scores: List<Score>): Int {
+    fun yieldRate(scores: List<Score>): Double {
         val totalProfit = scores.map { it -> it.prize }
             .reduce { acc, score -> acc + score }
-        return totalProfit / purchaseAmount
+        return totalProfit / purchaseAmount.toDouble()
     }
 }
